+++ src/main/java/org/acme/dto/PagamentoDTO.java
@@ -6,8 +6,8 @@ import java.time.LocalDate;
 
 public record PagamentoDTO(
         @NotNull(message = "O valor total deve ser maior que zero") Double valorTotal,
-        @NotBlank(message = "O método de pagamento não pode ser nulo") Integer idMetodo,
-        @NotBlank(message = "O status do pagamento não pode ser nulo") Integer idStatus,
+        @NotBlank(message = "O método de pagamento não pode ser nulo") Long idMetodo,
+        @NotBlank(message = "O status do pagamento não pode ser nulo") Long idStatus,
         @NotNull(message = "A data de pagamento não pode ser nula") LocalDate dataPagamento,
         Long idPedido
 ) {}
